% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\docType{methods}
\name{callVariants}
\alias{callVariants}
\alias{callVariants,SummarizedExperiment-method}
\title{Call mitochondrial heteroplasmic variants}
\usage{
callVariants(mitoSE, filterRefAllele = TRUE)

\S4method{callVariants}{SummarizedExperiment}(mitoSE, filterRefAllele = TRUE)
}
\arguments{
\item{mitoSE}{A RangedSummarizeExperiment initialized by
mgatk that will have variants called.}

\item{filterRefAllele}{Default = TRUE If the reference allele
is included in the mcols slot of the rowRanges of the mitoSE
object, then filter these when doing variant calling.}
}
\value{
A RangedSummarizedExperiment object with variants
called. This adds a 'variantFreq' slot to assays with the
per-sample variant allele frequency as well as meta data
to the rowRanges that summarizes the variants
}
\description{
\code{callVariants} takes a RangedSummarizedExperiment
and infers which allele, the frequence per sample,
and other meta data is approproriate for down-stream
analysis in mgatk.
}
\examples{

datafile <- paste0(system.file('extdata',package='mgatk'), "/mgatk_CML.txt.gz")
mitoSE <- importMito.txt(datafile)
vSE <- callVariants(mitoSE, filterRefAlle = FALSE)

}
