% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wDistFunction.R
\docType{methods}
\name{computeWeightedDistance}
\alias{computeWeightedDistance}
\alias{computeWeightedDistance,ANY-method}
\title{Compute weighted distance function.}
\usage{
computeWeightedDistance(x, y = NULL, method = "euc")

\S4method{computeWeightedDistance}{ANY}(x, y = NULL, method = "euc")
}
\arguments{
\item{x}{Data matrix of ratios (for mito: minor allele frequencies)}

\item{y}{Data matrix of weights (for mito: coverage)}

\item{method}{= "abs" The distance metric for the data.}
}
\value{
One or more outputs. See above
}
\description{
\code{computeWeightedDistance} takes a matrix of samples X features
}
\examples{
x <- matrix(rnorm(1000), nrow = 20) # 50 samples
y <- matrix(rpois(1000, 10), nrow = 20) # 50 samples
d <- computeWeightedDistance(x,y)

}
