% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blacklist.R
\docType{methods}
\name{filterKnownBlacklist}
\alias{filterKnownBlacklist}
\alias{filterKnownBlacklist,MultiAssayExperiment,character-method}
\title{Make a data.frame of reduced features for plotting}
\usage{
filterKnownBlacklist(mitoMAE, filter)

\S4method{filterKnownBlacklist}{MultiAssayExperiment,character}(mitoMAE, filter)
}
\arguments{
\item{mitoMAE}{A MultiAssayExperiment intialized by mgatk}

\item{filter}{A character signifing the position file to
filter mitochondrial variants from.}
}
\value{
Another mgatk object that is an S4 class
MultiAssayExperiment that removes variants
that are included in the specified blacklist.
}
\description{
\code{filterKnownBlacklist} takes a MultiAssayExperiment
used in mgatk and returns a subsetted object where mitochondrial
variant positions are removed based on pre-defined.
}
\details{
The name of the filter includes the genome build and the
experimental data source that was used to determine the
blacklist.

Current filters that are valid include:

'hg19_TF1'
}
\examples{
# Import an object
folder <-paste0(system.file('extdata',package='mgatk'),"/glioma/final")
mitoMAE <- importMito(folder)
dim(mitoMAE@ExperimentList[["coverage"]])

# Filter blacklist
mitoMAE2 <- filterKnownBlacklist(mitoMAE, "hg19_TF1")
dim(mitoMAE2@ExperimentList[["coverage"]])

}
\seealso{
filterKnownBlacklist
}
