% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\docType{methods}
\name{getMitoChr}
\alias{getMitoChr}
\alias{getMitoChr,MultiAssayExperiment-method}
\title{Get the mitochondrial chromosome from a MAE object}
\usage{
getMitoChr(mitoMAE)

\S4method{getMitoChr}{MultiAssayExperiment}(mitoMAE)
}
\arguments{
\item{mitoMAE}{A MultiAssayExperiment initilized by mgatk}
}
\value{
Character associated with the mitochondrial
chromosomes used in mgatk
}
\description{
\code{getMitoChr} takes a MultiAssayExperiment Object
and returns a character of the mitochondrial chromosome
name (e.g. "chrM" or "MT")
}
\examples{

folder <-paste0(system.file('extdata',package='mgatk'),"/glioma/final")
mitoMAE <- importMito(folder)
getMitoChr(mitoMAE)

}
