% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\docType{methods}
\name{importMito.explicit}
\alias{importMito.explicit}
\alias{importMito.explicit,character,character,character,character,character,character,character-method}
\title{Import data from plain text files explicitly point
to the requisite input files and render a RangedSummarized
Experiment.}
\usage{
importMito.explicit(Afile, Cfile, Gfile, Tfile, coverageFile, depthFile,
  referenceAlleleFile, mitoChr = "chrM", maxFreq = 0.5, minCoverage = 10)


  \S4method{importMito.explicit}{character,character,character,character,character,character,character}(Afile,
  Cfile, Gfile, Tfile, coverageFile, depthFile, referenceAlleleFile,
  mitoChr = "chrM", maxFreq = 0.5, minCoverage = 10)
}
\arguments{
\item{Afile}{Contains the position, sample, count of A alleles}

\item{Cfile}{Contains the position, sample, count of C alleles}

\item{Gfile}{Contains the position, sample, count of G alleles}

\item{Tfile}{Contains the position, sample, count of T alleles}

\item{coverageFile}{The filepath of a plain text or gzipped file
that contains data in a sparse matrix format of the position,
sample, and coverage. This will be imported and a
data object will be rendered that enables downstream
mgatk analyses.}

\item{depthFile}{The filepath of a plain text or gzipped file
that contains two columns indicating the sample name and the
second the mean coverage of the sample about the mitochondrial
genome}

\item{referenceAlleleFile}{The filepath of a plain text or gzipped file
that contains two columns indicating the position and designated
reference allele.}

\item{mitoChr}{Default = "chrM" The `seqname` of the mitochondrial
genome to be initialized in the RangedSummarizedExperiment object}

\item{maxFreq}{Default = 0.5 The maximum allele frequency for
a called variant}

\item{minCoverage}{Default = 10 The maximum number of total
reads for the variant to be carried forward}
}
\value{
An initialized mgatk object that is an S4 class
RangedSummarizedExperiment that includes variant
frequency and coverage as assay slots.
}
\description{
\code{importMito.explicit} takes a sparse matrix file of position,
sample, and count (non-zero) in some order as variant calls
(one file per allele) as well as
a separate file of the position and sample coverage and
produces a RangedSummarizedExperiment object that serves
as the backbone of the R interface for mgatk.
}
\details{
The filepath of a plain text or gzipped file
that contains data in a long matrix format of the position,
allele, sample, and coverage. This will be imported and a
data object will be rendered that enables downstream
mgatk analyses
}
\examples{

path <-paste0(system.file('extdata',package='mgatk'),"/glioma/final/")

Afile <-paste0(path, "mgatk.A.txt")
Cfile <-paste0(path, "mgatk.C.txt")
Gfile <-paste0(path, "mgatk.G.txt")
Tfile <-paste0(path, "mgatk.T.txt")

coverageFile <- paste0(path, "mgatk.coverage.txt")
depthFile <- paste0(path, "mgatk.depthTable.txt")
referenceAlleleFile <- paste0(path, "chrM_refAllele.txt")

mitoSE <- importMito.explicit(Afile, Cfile, Gfile, Tfile,
  coverageFile, depthFile, referenceAlleleFile)
dim(mitoSE)

}
