% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\docType{methods}
\name{importMito}
\alias{importMito}
\alias{importMito,character-method}
\title{Import data from folder output of mgatk python run.}
\usage{
importMito(folder, ...)

\S4method{importMito}{character}(folder)
}
\arguments{
\item{folder}{Filepath to folder (likely ending in 'final')
that contains all the vital input / output files for mgatk}

\item{...}{Additional parameters to pass to the
importMito.explict function}
}
\value{
An initialized mgatk object that is an S4 class
RangedSummarizedExperiment that includes variant
frequency and coverage as assay slots.
}
\description{
\code{importMito} takes a path to a folder that contains
all the requisitive input files for setting up an analysis
in mgatk. This is typically the `final` folder when running
the python mgatk utility.
}
\examples{
folder <-paste0(system.file('extdata',package='mgatk'),"/glioma/final")

mitoSE <- importMito(folder)
dim(mitoSE)

}
\seealso{
importMito.explicit
}
