% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\docType{methods}
\name{mitoFeatureReduction}
\alias{mitoFeatureReduction}
\alias{mitoFeatureReduction,SummarizedExperiment-method}
\title{Get a set of reduced features for plotting}
\usage{
mitoFeatureReduction(mitoSE, nPC, nTSNE, perplexity)

\S4method{mitoFeatureReduction}{SummarizedExperiment}(mitoSE, nPC = 10,
  nTSNE = 2, perplexity = 30)
}
\arguments{
\item{mitoSE}{A RangedSummarizedExperiment intialized by mgatk}

\item{nPC}{Default = 10; Number of principal components
to be returned and}

\item{nTSNE}{Default = 2; Number of tSNE coordinates to
be returned}

\item{perplexity}{Default = 30; Perplexity for tSNE computation}
}
\value{
A data.frame containing all of the
colData in the original SummarizedExperiment
as well as PCs and tSNE coordiantes
}
\description{
\code{mitoFeatureReduction} takes a RangedSummarizedExperiment
used in mgatk and returns a data.frame of reduced features
that can be used for plotting.
}
\details{
The data frame also includes all meta data that is
included in the SummarizedExperiment's colData.
}
\examples{
# Import an object
folder <-paste0(system.file('extdata',package='mgatk'),"/glioma/final")
mitoSE <- importMito(folder)

# Filter blacklist
mitoSE2 <- filterKnownBlacklist(mitoSE, "hg19_TF1")

df <- mitoFeatureReduction(mitoSE2, nPC = 3, nTSNE = 2, perplexity = 30)

}
