% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blacklist.R
\docType{methods}
\name{filterKnownBlacklist}
\alias{filterKnownBlacklist}
\alias{filterKnownBlacklist,SummarizedExperiment,character-method}
\title{Filter variants based on pre-defined blacklist files}
\usage{
filterKnownBlacklist(mitoSE, filter)

\S4method{filterKnownBlacklist}{SummarizedExperiment,character}(mitoSE, filter)
}
\arguments{
\item{mitoSE}{A RangedSummarizedExperiment intialized by mgatk}

\item{filter}{A character signifing the position file to
filter mitochondrial variants from.}
}
\value{
Another mgatk object that is an S4 class
RangedSummarizedExperiment that removes variants
that are included in the specified blacklist.
}
\description{
\code{filterKnownBlacklist} takes a RangedSummarizedExperiment
used in mgatk and returns a subsetted object where mitochondrial
variant positions are removed based on pre-defined.
}
\details{
The name of the filter includes the genome build and the
experimental data source that was used to determine the
blacklist.

Current filters that are valid include:

'hg19_TF1'
}
\examples{
# Import an object
folder <-paste0(system.file('extdata',package='mgatk'),"/glioma/final")
mitoSE <- importMito(folder)
dim(mitoSE)

# Filter blacklist
mitoSE2 <- filterKnownBlacklist(mitoSE, "hg19_TF1")
dim(mitoSE)

}
\seealso{
filterKnownBlacklist
}
