% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\docType{methods}
\name{importMito.txt}
\alias{importMito.txt}
\alias{importMito.txt,character,numeric-method}
\title{Import data from plain text file into}
\usage{
importMito.txt(datafile, pos = 1, allele = 2, sample = 3, coverage = 4,
  mito = "chrM")

\S4method{importMito.txt}{character,numeric}(datafile, pos = 1, allele = 2,
  sample = 3, coverage = 4, mito = "chrM")
}
\arguments{
\item{datafile}{The filepath of a plain text or gzipped file
that contains data in a long matrix format of the position,
allele, sample, and coverage. This will be imported and a
data object will be rendered that enables downstream
mgatk analyses}

\item{pos}{Default = 1 column index of the genomic position}

\item{allele}{Default = 2 column index of the allele (A/C/G/T)}

\item{sample}{Default = 3 column index of the sample name}

\item{coverage}{Default = 4 column index that shows the number
of reads for the sample}

\item{mito}{Default = "chrM" The `seqname` of the mitochondrial
genome to be initialized in the}
}
\value{
An initialized mgatk object that is an S4 class
RangedSummarizedExperiment
}
\description{
\code{importMito.txt} takes a long file of position, allele,
sample, and coverage (non-zero) in some order.
}
\examples{

datafile <- paste0(system.file('extdata',package='mgatk'), "/mgatk_CML.txt.gz")
mitoSE <- importMito.txt(datafile)

}
